+++ types/src/byte_string.rs
@@ -49,7 +49,7 @@ impl BinaryEncoder<ByteString> for ByteString {
         } else if len < -1 {
             error!("ByteString buf length is a negative number {}", len);
             Err(StatusCode::BadDecodingError)
-        } else if len as u32 > decoding_limits.max_byte_string_length {
+        } else if len as usize > decoding_limits.max_byte_string_length {
             error!("ByteString length {} exceeds decoding limit {}", len, decoding_limits.max_string_length);
             Err(StatusCode::BadDecodingError)
         } else {
